finish test coverage for player positive path

+++ src/main/java/com/dungeongroupfinder/controllers/PlayerController.java
@@ -2,18 +2,13 @@ package com.dungeongroupfinder.controllers;
 
 import com.dungeongroupfinder.entities.Player;
 import com.dungeongroupfinder.enums.Roles;
-import com.dungeongroupfinder.exceptions.PlayerNotFoundException;
 import com.dungeongroupfinder.services.PlayerService;
-import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
-import javax.xml.ws.Response;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/players")
@@ -32,42 +27,40 @@ public class PlayerController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
-    public void addPlayer(@RequestBody Player player) {
-        playerService.addPlayer(player);
+    public Player createPlayer(@RequestBody Player player) {
+        return playerService.createPlayer(player);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping("/{id}")
-    public ResponseEntity<Player> updatePlayer(@PathVariable int id, @RequestBody Player player) {
-        Optional<Player> playerOptional = playerService.getPlayerById(id);
-
-        if(!playerOptional.isPresent()) {
-            return ResponseEntity.notFound().build();
+    public Player updatePlayer(@PathVariable int id, @RequestBody Player player) {
+        Player foundPlayer = playerService.getPlayerById(id);
+        if(foundPlayer == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "A player with given id doesn't exist.");
         }
-
         player.setId(id);
-        playerService.addPlayer(player);
-        return ResponseEntity.noContent().build();
+        return playerService.createPlayer(player);
     }
 
     @GetMapping("/{id}")
-    public Player getPlayerById(@PathVariable int id) throws Exception {
-        Optional<Player> playerOptional = playerService.getPlayerById(id);
-        if(!playerOptional.isPresent()) {
-            throw new PlayerNotFoundException("id: " + id);
+    public Player getPlayerById(@PathVariable int id) {
+        Player foundPlayer = playerService.getPlayerById(id);
+        if(foundPlayer == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "A player with given id doesn't exist.");
         }
-        return playerOptional.get();
+        return foundPlayer;
     }
 
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PatchMapping("/{id}")
-    public void updatePlayerRoleById(@PathVariable int id, @RequestBody Roles role) {
-        playerService.updatePlayerRoleById(id, role);
-    }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     public void deletePlayer(@PathVariable int id) {
+        if(playerService.getPlayerById(id) == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "A player with given id doesn't exist.");
+        }
         playerService.deletePlayerById(id);
     }
 }
